App: decouple Instance/Configuration, move to client
- WIP toward APIs
- client and core will process their own configurations
- Instance won't (shouldn't) sit in client but will work for now
